Fix to get deployments from store once
